dagster_project/requirements.txt
dagster==1.4.13
dagster-webserver==1.4.13
dagster-postgres==0.20.13
elasticsearch==8.9.0
neo4j==5.13.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
requests>=2.28.1
pendulum==2.1.2
slack-sdk==3.26.1
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
elasticsearch-dsl>=8.9.0
neo4j-driver>=5.13.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyPDF2>=3.0.0
pdfminer.six>=20221105
tika>=2.6.0 